% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildExplainer.R
\name{buildExplainer}
\alias{buildExplainer}
\title{Step 1: Build an lightgbmExplainer}
\usage{
buildExplainer(lgb.model, trainingData, type = "binary", base_score = 0.5)
}
\arguments{
\item{lgb.model}{A trained lightgbm model}

\item{trainingData}{A lgb.Dataset of data used to train the model}

\item{type}{The objective function of the model - either "binary" (for binary:logistic) or "regression" (for reg:linear)}

\item{base_score}{Default 0.5. The base_score variable of the lightgbm model.}
}
\value{
The lightgbm Explainer for the model. This is a data table where each row is a leaf of a tree in the lightgbm model
 and each column is the impact of each feature on the prediction at the leaf.

 The leaf and tree columns uniquely identify the node.

 The sum of the other columns equals the prediction at the leaf (log-odds if binary response).

 The 'intercept' column is identical for all rows and is analogous to the intercept term in a linear / logistic regression.
}
\description{
This function outputs an lightgbmExplainer (a data table that stores the feature impact breakdown for each leaf of each tree in an lightgbm model). It is required as input into the explainPredictions and showWaterfall functions.
}
\examples{
library(lightgbm)
library(lightgbmExplainer)

lgb.train <- agaricus.train
lgb.dtrain <- lgb.Dataset(lgb.train$data, label = lgb.train$label)
lgb.params = list(objective = "binary")
lgb.model <- lgb.train(lgb.params, lgb.dtrain, 3)
lgb.trees <- lgb.model.dt.tree(lgb.model)
lgb.trees
library(lightgbmExplainer)
explainer = buildExplainer(xgb.model,xgb.train.data, type="binary", base_score = 0.5)
pred.breakdown = explainPredictions(xgb.model, explainer, xgb.test.data)

showWaterfall(xgb.model, explainer, xgb.test.data, test.data,  2, type = "binary")
showWaterfall(xgb.model, explainer, xgb.test.data, test.data,  8, type = "binary")
}
