% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explainPredictions.R
\name{explainPredictions}
\alias{explainPredictions}
\title{Step 2: Get multiple prediction breakdowns from a trained lightgbm model}
\usage{
explainPredictions(lgb.model, explainer, data)
}
\arguments{
\item{lgb.model}{A trained lightgbm model}

\item{explainer}{The output from the buildExplainer function, for this model}

\item{data}{A DMatrix of data to be explained}
}
\value{
A data table where each row is an observation in the data and each column is the impact of each feature on the prediction.

The sum of the row equals the prediction of the lightgbm model for this observation (log-odds if binary response).
}
\description{
This function outputs the feature impact breakdown of a set of predictions made using an lightgbm model.
}
\examples{
library(lightgbm) # v2.1.0 or above
library(lightgbmExplainer)

# Load Data
data(agaricus.train, package = "lightgbm")
# Train a model
lgb.dtrain <- lgb.Dataset(agaricus.train$data, label = agaricus.train$label)
lgb.params <- list(objective = "binary")
lgb.model <- lgb.train(lgb.params, lgb.dtrain, 5)
# Build Explainer
lgb.trees <- lgb.model.dt.tree(lgb.model) # First get a lgb tree
explainer <- buildExplainer(lgb.trees)
# compute contribution for each data point
pred.breakdown <- explainPredictions(lgb.model, explainer, agaricus.train$data)
# Show waterfall for the 8th observation
showWaterfall(lgb.model, explainer, lgb.dtrain, agaricus.train$data,  8, type = "binary")
}
