% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showWaterfall.R
\name{showWaterfall}
\alias{showWaterfall}
\title{Step 3: Get prediction breakdown and waterfall chart for a single row of data}
\usage{
showWaterfall(lgb.model, explainer, lgb.dtrain, lgb.train.data, id,
  type = "binary", threshold = 1e-04)
}
\arguments{
\item{lgb.model}{A trained lightgbm model}

\item{explainer}{The output from the buildExplainer function, for this model}

\item{lgb.dtrain}{The lgb.dtrain in which the row to be predicted is stored}

\item{lgb.train.data}{The matrix of data from which the lgb.dtrain was built}

\item{type}{The objective function of the model - either "binary" (for binary:logistic) or "regression" (for reg:linear)}

\item{threshold}{Default = 0.0001. The waterfall chart will group all variables with absolute impact less than the threshold into a variable called 'Other'}

\item{idx}{The row number of the data to be explained}
}
\value{
None
}
\description{
This function prints the feature impact breakdown for a single data row, and plots an accompanying waterfall chart.
}
