% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explainPredictions.R
\name{explainPredictions}
\alias{explainPredictions}
\title{Step 2: Get multiple prediction breakdowns from a trained lightgbm model}
\usage{
explainPredictions(lgb.model, explainer, data)
}
\arguments{
\item{lgb.model}{A trained lightgbm model}

\item{explainer}{The output from the buildExplainer function, for this model}

\item{data}{A DMatrix of data to be explained}
}
\value{
A data table where each row is an observation in the data and each column is the impact of each feature on the prediction.

The sum of the row equals the prediction of the lightgbm model for this observation (log-odds if binary response).
}
\description{
This function outputs the feature impact breakdown of a set of predictions made using an lightgbm model.
}
\examples{
library(lightgbm)
library(lightgbmExplainer)

set.seed(123)

data(agaricus.train, package='lightgbm')

X = as.matrix(agaricus.train$data)
y = agaricus.train$label

train_idx = 1:5000

train.data = X[train_idx,]
test.data = X[-train_idx,]

lgb.train.data <- lgb.DMatrix(train.data, label = y[train_idx])
lgb.test.data <- lgb.DMatrix(test.data)

param <- list(objective = "binary:logistic")
lgb.model <- lightgbm(param =param,  data = lgb.train.data, nrounds=3)

col_names = colnames(X)

pred.train = predict(lgb.model,X)
nodes.train = predict(lgb.model,X,predleaf =TRUE)
trees = lgb.model.dt.tree(col_names, model = lgb.model)

#### The lightgbm Explainer
explainer = buildExplainer(lgb.model,lgb.train.data, type="binary", base_score = 0.5)
pred.breakdown = explainPredictions(lgb.model, explainer, lgb.test.data)

showWaterfall(lgb.model, explainer, lgb.test.data, test.data,  2, type = "binary")
showWaterfall(lgb.model, explainer, lgb.test.data, test.data,  8, type = "binary")
}
